backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
gunicorn
daphne
channels>=4.0
channels-redis
celery
redis
psycopg2-binary
djangorestframework-simplejwt
numpy
psutil
python-dotenv
watchdog